Refactor settings into base settings, fixup modules
TODO: reset
