+++ EyeTrackApp/osc.py
@@ -2,11 +2,13 @@
 from pythonosc import udp_client
 from pythonosc import osc_server
 from pythonosc import dispatcher
+from config import EyeTrackConfig
 from utils.misc_utils import PlaySound,SND_FILENAME,SND_ASYNC
 import queue
 import threading
 from enum import IntEnum
 import time
+
 
 class EyeId(IntEnum):
     RIGHT = 0
@@ -14,7 +16,7 @@ class EyeId(IntEnum):
     BOTH = 2
     SETTINGS = 3
     ALGOSETTINGS = 4
-from config import EyeTrackConfig
+
 
 def eyelid_transformer(self,eye_blink):
     if self.config.osc_invert_eye_close:
